fix: types, minor issues
