+++ src/services/cd-ng/index.tsx
@@ -10872,7 +10872,6 @@ export interface ServiceSpec {
   configFiles?: ConfigFileWrapper[]
   manifests?: ManifestConfigWrapper[]
   variables?: NGVariable[]
-  startupScript?: any
 }
 
 export interface ServiceUsageDTO {
